+++ app/Config/Autoload.php
@@ -53,6 +53,7 @@ class Autoload extends AutoloadConfig
         'App\Modules\Iam'       => APPPATH . 'Modules/Iam',
         'App\Modules\System'    => APPPATH . 'Modules/System',
         'App\Modules\Catalog'  => APPPATH . 'Modules/Catalog',
+        'App\Modules\Cms'  => APPPATH . 'Modules/Cms',
     ];
 
     /**